ci: check for noeol files
Yes, this should have a ! to negate the exit code.

But for some reason, when run in bash -c, the exit code is inverted.
Why? I can't possibly say; this is what works.

It's not ideal that it doesn't show which files, but there is no
principled way to check for noeol files, surprisingly enough, and this
is the best we can do consistently.